Upload env vars during next.js deployment.
