+++ src/commands/deploy/genezio.ts
@@ -24,7 +24,6 @@ import {
     directoryContainsHtmlFiles,
     deleteFolder,
     getBundleFolderSizeLimit,
-    readEnvironmentVariablesFile,
     zipFile,
 } from "../../utils/file.js";
 import { printAdaptiveLog, debugLogger, doAdaptiveLogAction } from "../../utils/logging.js";
@@ -42,10 +41,7 @@ import {
 import { CloudProviderIdentifier } from "../../models/cloudProviderIdentifier.js";
 import { GenezioDeployOptions } from "../../models/commandOptions.js";
 import { GenezioTelemetry, TelemetryEventTypes } from "../../telemetry/telemetry.js";
-import { setEnvironmentVariables } from "../../requests/setEnvironmentVariables.js";
 import colors from "colors";
-import { getEnvironmentVariables } from "../../requests/getEnvironmentVariables.js";
-import { getProjectEnvFromProject } from "../../requests/getProjectInfo.js";
 import { Status } from "../../requests/models.js";
 import { bundle } from "../../bundlers/utils.js";
 import {
@@ -72,6 +68,7 @@ import { uploadContentToS3 } from "../../requests/uploadContentToS3.js";
 import {
     getOrCreateDatabase,
     getOrCreateEmptyProject,
+    handleEnvVars,
     processYamlEnvironmentVariables,
 } from "./utils.js";
 
@@ -327,6 +324,8 @@ export async function genezioDeploy(options: GenezioDeployOptions) {
     for (const frontendUrl of frontendUrls) {
         log.info(colors.cyan(`Frontend URL: ${frontendUrl}`));
     }
+
+    return deployClassesResult;
 }
 
 export async function deployClasses(
@@ -530,104 +529,11 @@ export async function deployClasses(
     const projectId = result.projectId;
     const projectEnvId = result.projectEnvId;
     if (projectId) {
-        // Deploy environment variables if --upload-env is true
-        if (options.env) {
-            const envFile = path.join(process.cwd(), options.env);
-            debugLogger.debug(`Loading environment variables from ${envFile}.`);
-
-            if (!(await fileExists(envFile))) {
-                // There is no need to exit the process here, as the project has been deployed
-                log.error(`File ${envFile} does not exists. Please provide the correct path.`);
-                await GenezioTelemetry.sendEvent({
-                    eventType: TelemetryEventTypes.GENEZIO_DEPLOY_ERROR,
-                    errorTrace: `File ${envFile} does not exists`,
-                });
-            } else {
-                // Read environment variables from .env file
-                const envVars = await readEnvironmentVariablesFile(envFile);
-                const projectEnv = await getProjectEnvFromProject(projectId, options.stage);
-
-                if (!projectEnv) {
-                    throw new UserError("Project environment not found.");
-                }
-
-                // Upload environment variables to the project
-                await setEnvironmentVariables(projectId, projectEnv.id, envVars)
-                    .then(async () => {
-                        debugLogger.debug(
-                            `Environment variables from ${envFile} uploaded to project ${projectId}`,
-                        );
-                        log.info(
-                            `The environment variables were uploaded to the project successfully.`,
-                        );
-                        await GenezioTelemetry.sendEvent({
-                            eventType: TelemetryEventTypes.GENEZIO_DEPLOY_LOAD_ENV_VARS,
-                        });
-                    })
-                    .catch(async (error: AxiosError) => {
-                        log.error(`Loading environment variables failed with: ${error.message}`);
-                        log.error(
-                            `Try to set the environment variables using the dashboard ${colors.cyan(
-                                DASHBOARD_URL,
-                            )}`,
-                        );
-                        await GenezioTelemetry.sendEvent({
-                            eventType: TelemetryEventTypes.GENEZIO_DEPLOY_ERROR,
-                            errorTrace: error.toString(),
-                        });
-                    });
-            }
-        } else {
-            const cwd = projectConfiguration.workspace?.backend
-                ? path.resolve(projectConfiguration.workspace.backend)
-                : process.cwd();
-            const envFile = path.join(cwd, ".env");
-            if (await fileExists(envFile)) {
-                // read envVars from file
-                const envVars = await readEnvironmentVariablesFile(envFile);
-                const projectEnv = await getProjectEnvFromProject(projectId, options.stage);
-
-                if (!projectEnv) {
-                    throw new UserError("Project environment not found.");
-                }
-
-                // get remoteEnvVars from project
-                const remoteEnvVars = await getEnvironmentVariables(projectId, projectEnv.id);
-
-                // check if all envVars from file are in remoteEnvVars
-                const missingEnvVars = envVars.filter(
-                    (envVar) =>
-                        !remoteEnvVars.find((remoteEnvVar) => remoteEnvVar.name === envVar.name),
-                );
-
-                // Print missing env vars
-                if (missingEnvVars.length > 0) {
-                    log.info(
-                        `${colors.yellow(
-                            "Warning: The following environment variables are not set on your project: ",
-                        )}`,
-                    );
-                    missingEnvVars.forEach((envVar) => {
-                        log.info(`${colors.yellow(envVar.name)}`);
-                    });
-
-                    const relativeEnvFilePath = path.join(
-                        ".",
-                        path.relative(path.resolve(process.cwd()), path.resolve(envFile)),
-                    );
-
-                    log.info("");
-                    log.info(
-                        `${colors.yellow("Go to the dashboard ")}${colors.cyan(
-                            DASHBOARD_URL,
-                        )} ${colors.yellow(
-                            "to set your environment variables or run ",
-                        )} ${colors.cyan(`genezio deploy --env ${relativeEnvFilePath}`)}`,
-                    );
-                    log.info("");
-                }
-            }
-        }
+        // Deploy environment variables if --env is true
+        const cwd = projectConfiguration.workspace?.backend
+            ? path.resolve(projectConfiguration.workspace.backend)
+            : process.cwd();
+        handleEnvVars(options.env, projectId, projectEnvId, cwd);
 
         return {
             projectId: projectId,
